Update README.md
PCM-TESTING is a python-based application for XII th grade science stream(mainly for PCM) students to test to themselves for their Boards Examinations.

To make this code run, you have to makes some changes according to your system configuration. Firstly, you have downlaod the folder 'qbank' and save it into C-Drive users and after that change the directory mapping of the coding with the directory of the folder'qbank' of your system. And, Make it Run